Refactor: pianoroll is now a module
